combined 'magic' and 'spellList' into 'Magic' module
