+++ client/www/views/components/spells/Magic.js
@@ -1,0 +1,40 @@
+var Magic = {
+  setPower: function() {return Math.floor(Math.random() * 10 + 1);},
+  setCrit: function() {
+    var roll = Math.floor(Math.random() * 20 + 1);
+    var crit = null;
+    if (roll > 17) return 1;
+    if (roll < 3) return -1;
+    return 0;
+  },
+  setTime: function() {return new Date().getTime();},
+
+  castSpell: function(attack, power, crit, timeShift) {
+    var spell = {
+      attackId: attack.attackId,
+      power: power ? power : this.setPower(),
+      crit: crit !== null ? crit : this.setCrit(),
+      time: this.setTime() + timeShift,
+      spellName: attack.spellName
+    };
+    return spell;
+  },
+  spellList: {
+    attacks: [
+      { name: 'Magic Missile', type: 'Attack', target: 'foe', role: 'attack', afinity: 'basic', cost: 5},
+      {name: 'Water Coffin', type: 'Attack', target: 'foe', role: 'attack', afinity: 'water', cost: 7},
+      {name: 'Wind Swords', type: 'Attack', target: 'foe', role: 'attack', afinity:'air', cost: 7},
+    ],
+    counters: [
+
+      { name: 'Warp spacetime', icon: 'ion-android-favorite-outline', type: 'Perry', target: 'foe', role: 'perry', afinity: 'basic', cost: 5 },
+      { name: 'Mystical Judo', icon: 'ion-ios-plus-outline', type: 'Repost', target: 'foe', role: 'repost', afinity: 'basic', cost: 6 },
+    ],
+    enhancers: [
+      {name: 'Heal', icon: 'ion-heart', type: 'recovery', target: 'caster', role: 'heal', afinity: 'basic', cost: 5, power: 5},
+      {name: 'Force Armor', icon: 'ion-ios-plus-outline', type: 'buff', target: 'caster', role:'buff', afinity:'basic', cost: 5, duration: 15},
+    ]
+  }
+};
+
+module.exports = Magic;
